Initial commit for DroidLedger
DroidLedger is a hybrid mobile/web app front-end for
ledger-cli. Currently it is in a mockup stage and doesn't offer any
integration with any existing backend services.

The idea is to use this in the future with a server capable of handling
RESTful api calls for reading and adding information regarding
transactions in a way that is friendly to the existing ledger-cli
file format.